TEST: cover log.log_gpu_memory branches and move to test_log.py
CI's coverage check (>=90% on diff) was failing on the log_gpu_memory
try/except added in c98af28: lines 70-74 of log.py weren't reached.

Two issues:

1. The TestGpuMemoryLogging class lived in test_lifecycle.py, which
   does pytest.importorskip on the GCG train module. CI installs the
   'all' extra but not 'gcg', so ml_collections is missing and the
   train import fails, skipping the whole test_lifecycle.py module --
   including the GPU memory tests, even though they only need stdlib.
   Moved them into test_log.py (which only importorskips the log
   module, all stdlib) so they actually run in CI.

2. The new test_log_gpu_memory_swallows_nvidia_smi_failure exercises
   the except branch (lines 73-74) that the old success-only test
   never hit. log_gpu_memory must swallow nvidia-smi failures so the
   training loop never crashes when run on a host without nvidia-smi.
